SRV-2198: Gateway codegen: 'workflow-parts' endpoint restructured
(i.e. endpoint to copy and paste parts of workflows)
In order to be more consistent with the other endpoints.
What has been done: the workflow-id refering to the workflow to
paste/copy the parts to/from has changed from a property (of
WorkfowParts) to a path-parameter.

SRV-2198 (Gateway API: Fix duplicated endpoints and inconsistency with
'parts'-endpoint)